refactor(model): breakdown the model and convert to ts
